Make openapsama_current_basal_safety_multiplier Double instead of Int

+++ app/src/main/java/info/nightscout/androidaps/plugins/ConstraintsSafety/SafetyPlugin.java
@@ -130,7 +130,7 @@ public class SafetyPlugin implements PluginBase, ConstraintsInterface {
         if (profile == null) return absoluteRate;
         if (absoluteRate < 0) absoluteRate = 0d;
 
-        Integer maxBasalMult = SP.getInt("openapsama_current_basal_safety_multiplier", 4);
+        Double maxBasalMult = SP.getDouble("openapsama_current_basal_safety_multiplier", 4d);
         Integer maxBasalFromDaily = SP.getInt("openapsama_max_daily_safety_multiplier", 3);
         // Check percentRate but absolute rate too, because we know real current basal in pump
         Double origRate = absoluteRate;
@@ -168,7 +168,7 @@ public class SafetyPlugin implements PluginBase, ConstraintsInterface {
 
         if (absoluteRate < 0) absoluteRate = 0d;
 
-        Integer maxBasalMult = SP.getInt("openapsama_current_basal_safety_multiplier", 4);
+        Double maxBasalMult = SP.getDouble("openapsama_current_basal_safety_multiplier", 4d);
         Integer maxBasalFromDaily = SP.getInt("openapsama_max_daily_safety_multiplier", 3);
         // Check percentRate but absolute rate too, because we know real current basal in pump
         Double origRate = absoluteRate;
